Other university data can be processed
Other university data can be processed

ID of psychological test result differ from real ID.
And so, now, convert the ID.

Drug tables does not involve ID but key of ID exists.
And so, drug table is not default value.

README was modified
  LibreOffice is now recommended

Class to modify ID was wrote.
It is needed for convert ID of psychological test.